backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-multipart
google-generativeai
pydantic
